Update prop naming for Booleans in Availability Label component & link
Props and variables that hold Boolean values should be named according
to the common standard, which was not the case for Availability Label
component and link utility.

Availability Label story was set up with a flaw, where the link prop
was defined on the label in "selected" state. This now reflects reality,
and the link prop is undefined for this state.